Start multi-version support
