+++ src/v1_2/operand/file_offset.rs
@@ -2,7 +2,7 @@
 use crate::test_tools::test_item;
 use crate::{
     codec::{Codec, StdError, WithOffset, WithSize},
-    varint,
+    v1_2::varint,
 };
 #[cfg(test)]
 use hex_literal::hex;
